Refacto

+++ libraries/Vasco_IOT/src/data/Filter.cpp
@@ -1,38 +1,38 @@
 #include "Filter.h"
 
-int Filter::_isometric(float value, float maximum, float minimum, int height, int offset) 
+float Filter::_isometric(float value, float maximum, float minimum, int height, int offset)
 {
-  return int((1 - ((float) (value - minimum) / (float) (maximum - minimum))) * height) + offset;
+  return (1 - ((float)(value - minimum) / (float)(maximum - minimum))) * height + offset;
 }
 
-float Filter::_smooth(float value, int factor) 
+float Filter::_smooth(float value, int factor)
 {
   if (factor >= _BUFFER_MAX_SIZE) {
     Serial.print(F("Can't smmoth, factor must be < "));
     Serial.println(_BUFFER_MAX_SIZE);
-    
+
     return 0;
   }
-  
+
   // moving average algorithm
   _smoothTotal -= _smoothBuffer[_smoothIndex];
   _smoothBuffer[_smoothIndex] = value;
   _smoothTotal += _smoothBuffer[_smoothIndex];
-  
+
   _smoothIndex++;
   if (_smoothIndex >= factor) {
     _smoothIndex = 0;
   }
 
-  return _smoothTotal / (float) factor;
+  return _smoothTotal / (float)factor;
 }
 
-float Filter::_maximum(float value, int factor) 
+float Filter::_maximum(float value, int factor)
 {
   if (factor >= _BUFFER_MAX_SIZE) {
     Serial.print(F("Can't smmoth, factor must be < "));
     Serial.println(_BUFFER_MAX_SIZE);
-    
+
     return 0;
   }
 
@@ -53,20 +53,20 @@ float Filter::_maximum(float value, int factor)
   return maximum;
 }
 
-float Filter::_inertia(float value, int factor) 
+float Filter::_inertia(float value, int factor)
 {
   if (factor >= _BUFFER_MAX_SIZE) {
     Serial.print(F("Can't smmoth, factor must be < "));
     Serial.println(_BUFFER_MAX_SIZE);
-    
+
     return 0;
   }
 
-  // inertia algorithm      
+  // inertia algorithm
   if (value > _inertiaValue) {
-    _inertiaValue += (value - _inertiaValue) / (float) factor;
+    _inertiaValue += (value - _inertiaValue) / (float)factor;
   } else if (value < _inertiaValue) {
-    _inertiaValue -= (_inertiaValue - value) / (float) factor;
+    _inertiaValue -= (_inertiaValue - value) / (float)factor;
   }
 
   return _inertiaValue;
